Add text replacement to new JSON interface
